+++ packages/uipath_langchain_client/src/uipath_langchain_client/clients/fireworks/embeddings.py
@@ -29,13 +29,11 @@ class UiPathFireworksEmbeddings(UiPathBaseLLMClient, FireworksEmbeddings):
     def setup_uipath_client(self) -> Self:
         self.client = OpenAI(
             api_key="PLACEHOLDER",
-            timeout=None,  # handled by the UiPath client
             max_retries=0,  # handled by the UiPath client
             http_client=self.uipath_sync_client,
         )
         self.async_client = AsyncOpenAI(
             api_key="PLACEHOLDER",
-            timeout=None,  # handled by the UiPath client
             max_retries=0,  # handled by the UiPath client
             http_client=self.uipath_async_client,
         )