+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `uipath_llm_client` (core package) will be documented in this file.
 
+## [1.2.1] - 2026-02-18
+
+### Fix
+- TImeout fixes, change typing from int to float
+- remove timeout=None from all clients -> caused overriding the default timeout set up on the UiPathHttpxClient
+
 ## [1.2.0] - 2026-02-18
 
 ### Stable release